fix: move tests to layout_templ_test file
